Add files via upload
Git wouldnt let me pull your comments so i just copypasted your comments in and then tinkered with it a bit so it now works more! sum has been changed to sumarray, and it now populates grid with the appropriate changes for the next step. 
it does not update the grid graphic with the changes, and it jams after one loop but its closer!